testing configs

+++ docker-makefiles/clean_configs.sh
@@ -51,10 +51,16 @@ SNAPSHOT_FILE="${SNAPSHOT_DIR}/formatted_node_configs.env"
 # -------- Generate Docker Compose File --------
 DOCKER_DIR="docker-makefiles/docker-compose-files"
 DOCKER_FILE="${DOCKER_DIR}/${NODE_CONFIGS}-docker-compose.yaml"
+NODE_NAME_FILE="$(dirname "${SNAPSHOT_FILE}")/NODE_NAME.txt"
 
 # -------- Remove files --------
-rm -rf ${DOCKER_FILE} "${SNAPSHOT_FILE}"
 
-if [[ -d "${DOCKER_DIR}" && -z "$(ls -A "${DOCKER_DIR}")" ]]; then
-  rm -rf "${DOCKER_DIR}"
-fi
+printf "Warning: this will fully reset this node's deployment configuration.\nThe node will be treated as new on next startup and assigned a new identity.\nProceed? [y/N]: "
+read -r CONFIRM_DELETE
+
+if [[ "${CONFIRM_DELETE}" == "y" || "${CONFIRM_DELETE}" == "Y" ]]; then
+  rm -rf "${DOCKER_FILE}" "${SNAPSHOT_FILE}" "${NODE_NAME_FILE}"
+  if [[ -d "${DOCKER_DIR}" && -z "$(ls -A "${DOCKER_DIR}")" ]]; then
+    rm -rf "${DOCKER_DIR}"
+  fi
+else echo "Deletion cancelled." ; fi
